Refactor action name management
This commit refactors the way action named are managed. It becomes more cumbersone create a new action and add it to the saga module, since it will require to duplicate the string that acts as unique identifier for the action in the module.

This commit is required for the introducion of the Async/Sync actions extension that will be implemented in the next commits. The extra complexity will be managed by this extension and the end result will be better than the current approach.

Unfortunately it seems there is not way to both achieve Async/Sync actions and also avoid this refactor. The assumption here (validate by our usage of the same library in JS) is that we won’t ever use plain actions but we will always go through this new system that will be implemented in the next commits