Refactor SideEffectDependencyContainer
This commit adds a new way to handle dependency container.

The most important change is that the dependency container is not instantiated anymore on each side effect. While the purpose of preventing the state in the container, it leads to very weird and hard to maintein usages of the managers (in the best case, singletons) to work with the Apple framework. This defeats entirely the purpose of the instantiation every time and therefore we decided to remove it.

The second big change is that the container doesn’t receive the state anymore. We have realized that managers/dependencies shouldn’t have access to the state but they should use actions.

Third and last change is the class enforcement of the container. We have realized that without creating the container every time we may have issues with structs. In order to prevent these issues, we now enforce that the contains should be a class